+++ components/engine_traits/Cargo.toml
@@ -13,6 +13,7 @@ error_code = { path = "../error_code", default-features = false }
 fail = "0.5"
 file_system = { path = "../file_system", default-features = false }
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
+lazy_static = "1.3"
 log_wrappers = { path = "../log_wrappers" }
 protobuf = "2"
 raft = { version = "0.7.0", default-features = false, features = ["protobuf-codec"] }
